requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
anthropic==0.20.0
supabase==2.3.0
python-dotenv==1.0.0
schedule==1.2.1
pandas==2.1.4
loguru==0.7.2
